refactor(twitter): share query id resolution
